requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
geopandas
numpy
jsonschema
sentence-transformers
ruff
